+++ CHANGELOG.md
@@ -4,14 +4,15 @@
   partitioning brick.
 * Added a file type detection module.
 * Updated `partition_html` and `partition_eml` to support file-like objects in 'rb' mode.
-
-## 0.3.6-dev1
-
 * Cleaning brick for removing ordered bullets `clean_ordered_bullets`.
 * Extract brick method for ordered bullets `extract_ordered_bullets`.
 * Test for `clean_ordered_bullets`.
 * Test for `extract_ordered_bullets`.
 * Added `partition_docx` for pre-processing Word Documents.
+* Added new REGEX patterns to extract email header information
+* Added new functions to extract header information `parse_received_data` and `partition_header`
+* Added new function to parse plain text files `partition_text`
+* Added new cleaners functions `extract_ip_address`, `extract_ip_address_name`, `extract_mapi_id`, `extract_datetimetz`
 
 ## 0.3.5
 
@@ -25,6 +26,7 @@
 * Add new function `extract_attachment_info` that extracts and decode the attachment
 of an email.
 * Staging brick to convert a list of `Element`s to a `pandas` dataframe.
+* Add plain text functionality to `partition_email`
 
 ## 0.3.4